Extract DynamicKeyStore
To encapsulate a Key Store that can add new certificates to itself
