+++ java8/src/main/java/com/github/tomakehurst/wiremock/http/ssl/CertificateGeneratingX509ExtendedKeyManager.java
@@ -18,13 +18,9 @@ import java.security.NoSuchProviderException;
 import java.security.Principal;
 import java.security.PrivateKey;
 import java.security.SignatureException;
-import java.security.UnrecoverableKeyException;
-import java.security.cert.Certificate;
 import java.security.cert.CertificateException;
 import java.security.cert.X509Certificate;
 import java.util.ArrayList;
-import java.util.Collections;
-import java.util.Enumeration;
 import java.util.List;
 
 import static java.util.Collections.emptyList;
@@ -32,66 +28,28 @@ import static java.util.Collections.emptyList;
 @SuppressWarnings("sunapi")
 public class CertificateGeneratingX509ExtendedKeyManager extends DelegatingX509ExtendedKeyManager {
 
-    private final KeyStore keyStore;
-    private final char[] password;
-    private final CertificateAuthority existingCertificateAuthority;
+    private final DynamicKeyStore dynamicKeyStore;
 
     public CertificateGeneratingX509ExtendedKeyManager(X509ExtendedKeyManager keyManager, KeyStore keyStore, char[] keyPassword) {
         super(keyManager);
-        this.keyStore = keyStore;
-        this.password = keyPassword;
-        existingCertificateAuthority = findExistingCertificateAuthority();
+        dynamicKeyStore = new DynamicKeyStore(keyStore, keyPassword);
     }
 
     @Override
     public PrivateKey getPrivateKey(String alias) {
         PrivateKey original = super.getPrivateKey(alias);
-        return original == null ? getDynamicPrivateKey(alias) : original;
-    }
-
-    private PrivateKey getDynamicPrivateKey(String alias) {
-        try {
-            return (PrivateKey) keyStore.getKey(alias, password);
-        } catch (KeyStoreException | NoSuchAlgorithmException | UnrecoverableKeyException e) {
-            return null;
-        }
+        return original == null ? dynamicKeyStore.getPrivateKey(alias) : original;
     }
 
     @Override
     public X509Certificate[] getCertificateChain(String alias) {
         X509Certificate[] original = super.getCertificateChain(alias);
         if (original == null) {
-            return getDynamicCertificateChain(alias);
+            return dynamicKeyStore.getCertificateChain(alias);
         } else {
             return original;
         }
     }
-
-    private X509Certificate[] getDynamicCertificateChain(String alias) {
-        try {
-            Certificate[] fromKeyStore = keyStore.getCertificateChain(alias);
-            if (fromKeyStore != null && areX509Certificates(fromKeyStore)) {
-                return convertToX509(fromKeyStore);
-            } else {
-                return null;
-            }
-        } catch (KeyStoreException e) {
-            return null;
-        }
-    }
-
-    private static boolean areX509Certificates(Certificate[] fromKeyStore) {
-        return fromKeyStore.length == 0 || fromKeyStore[0] instanceof X509Certificate;
-    }
-
-    private static X509Certificate[] convertToX509(Certificate[] fromKeyStore) {
-        X509Certificate[] result = new X509Certificate[fromKeyStore.length];
-        for (int i = 0; i < fromKeyStore.length; i++) {
-            result[i] = (X509Certificate) fromKeyStore[i];
-        }
-        return result;
-    }
-
 
     @Override
     public String chooseServerAlias(String keyType, Principal[] issuers, Socket socket) {
@@ -154,7 +112,7 @@ public class CertificateGeneratingX509ExtendedKeyManager extends DelegatingX509E
      * @param handshakeSession nullable
      */
     private String tryToChooseServerAlias(String keyType, String defaultAlias, ExtendedSSLSession handshakeSession) {
-        if (defaultAlias != null && handshakeSession != null && existingCertificateAuthority != null) {
+        if (defaultAlias != null && handshakeSession != null && dynamicKeyStore.hasCertificateAuthority()) {
             return chooseServerAlias(keyType, defaultAlias, handshakeSession);
         } else {
             return defaultAlias;
@@ -206,59 +164,14 @@ public class CertificateGeneratingX509ExtendedKeyManager extends DelegatingX509E
             return defaultAlias;
         } else {
             try {
-                return generateCertificate(keyType, requestedServerNames.get(0));
+                SNIHostName requestedServerName = requestedServerNames.get(0);
+                dynamicKeyStore.generateCertificate(keyType, requestedServerName);
+                return requestedServerName.getAsciiName();
             } catch (KeyStoreException | CertificateException | NoSuchAlgorithmException | NoSuchProviderException | InvalidKeyException | SignatureException e) {
                 // TODO log?
                 return defaultAlias;
             }
         }
-    }
-
-    /**
-     * @param keyType non null, guaranteed to be valid
-     * @param requestedServerName non null
-     * @return an alias to a new private key & certificate for the first requested server name
-     */
-    private String generateCertificate(
-        String keyType,
-        SNIHostName requestedServerName
-    ) throws CertificateException, NoSuchAlgorithmException, SignatureException, NoSuchProviderException, InvalidKeyException, KeyStoreException {
-
-        CertChainAndKey newCertChainAndKey = existingCertificateAuthority.generateCertificate(keyType, requestedServerName);
-
-        String requestedNameString = requestedServerName.getAsciiName();
-        keyStore.setKeyEntry(requestedNameString, newCertChainAndKey.key, password, newCertChainAndKey.certificateChain);
-        return requestedNameString;
-    }
-
-    private CertificateAuthority findExistingCertificateAuthority() {
-        Enumeration<String> aliases;
-        try {
-            aliases = keyStore.aliases();
-        } catch (KeyStoreException e) {
-            aliases = Collections.emptyEnumeration();
-        }
-        while (aliases.hasMoreElements()) {
-            String alias = aliases.nextElement();
-            CertificateAuthority key = getCertChainAndKey(alias);
-            if (key != null) return key;
-        }
-        return null;
-    }
-
-    private CertificateAuthority getCertChainAndKey(String alias) {
-        X509Certificate[] chain = getCertificateChain(alias);
-        PrivateKey key = getPrivateKey(alias);
-        if (isCertificateAuthority(chain[0]) && key != null) {
-            return new CertificateAuthority(chain, key);
-        } else {
-            return null;
-        }
-    }
-
-    private static boolean isCertificateAuthority(X509Certificate certificate) {
-        boolean[] keyUsage = certificate.getKeyUsage();
-        return keyUsage != null && keyUsage.length > 5 && keyUsage[5];
     }
 
     private static boolean matches(X509Certificate x509Certificate, List<SNIHostName> requestedServerNames) {
